requirements.txt
openai>=1.50.0
anthropic>=0.40.0
litellm>=1.50.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
scikit-learn>=1.4.0
torch>=2.2.0
numpy>=1.26.0
plotly>=5.20.0
gradio>=5.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
requests>=2.31.0
tqdm>=4.66.0
pydantic>=2.5.0
modal>=0.66.0
huggingface-hub>=0.25.0
datasets>=2.20.0
